Rename package names to `org.fossify.*`
