--- app/src/main/kotlin/com/simplemobiletools/draw/pro/App.kt
@@ -1,11 +1,0 @@
-package com.simplemobiletools.draw.pro
-
-import android.app.Application
-import com.simplemobiletools.commons.extensions.checkUseEnglish
-
-class App : Application() {
-    override fun onCreate() {
-        super.onCreate()
-        checkUseEnglish()
-    }
-}
