--- app/src/main/kotlin/com/simplemobiletools/draw/pro/activities/SplashActivity.kt
@@ -1,11 +1,0 @@
-package com.simplemobiletools.draw.pro.activities
-
-import android.content.Intent
-import com.simplemobiletools.commons.activities.BaseSplashActivity
-
-class SplashActivity : BaseSplashActivity() {
-    override fun initActivity() {
-        startActivity(Intent(this, MainActivity::class.java))
-        finish()
-    }
-}
