--- app/src/main/kotlin/com/simplemobiletools/draw/pro/extensions/Context.kt
@@ -1,6 +1,0 @@
-package com.simplemobiletools.draw.pro.extensions
-
-import android.content.Context
-import com.simplemobiletools.draw.pro.helpers.Config
-
-val Context.config: Config get() = Config.newInstance(applicationContext)
