--- app/src/main/kotlin/com/simplemobiletools/draw/pro/helpers/Constants.kt
@@ -1,14 +1,0 @@
-package com.simplemobiletools.draw.pro.helpers
-
-const val BRUSH_COLOR = "brush_color"
-const val CANVAS_BACKGROUND_COLOR = "canvas_background_color"
-const val SHOW_BRUSH_SIZE = "show_brush_size"
-const val BRUSH_SIZE = "brush_size_2"
-const val LAST_SAVE_FOLDER = "last_save_folder"
-const val LAST_SAVE_EXTENSION = "last_save_extension"
-const val ALLOW_ZOOMING_CANVAS = "allow_zooming_canvas"
-const val FORCE_PORTRAIT_MODE = "force_portrait_mode"
-
-const val PNG = "png"
-const val SVG = "svg"
-const val JPG = "jpg"
