--- app/src/main/kotlin/com/simplemobiletools/draw/pro/interfaces/CanvasListener.kt
@@ -1,7 +1,0 @@
-package com.simplemobiletools.draw.pro.interfaces
-
-interface CanvasListener {
-    fun toggleUndoVisibility(visible: Boolean)
-
-    fun toggleRedoVisibility(visible: Boolean)
-}
