+++ app/src/main/kotlin/org/fossify/draw/App.kt
@@ -1,0 +1,11 @@
+package org.fossify.draw
+
+import android.app.Application
+import org.fossify.commons.extensions.checkUseEnglish
+
+class App : Application() {
+    override fun onCreate() {
+        super.onCreate()
+        checkUseEnglish()
+    }
+}
