+++ app/src/main/kotlin/org/fossify/draw/actions/Action.kt
@@ -1,0 +1,11 @@
+package org.fossify.draw.actions
+
+import android.graphics.Path
+import java.io.Serializable
+import java.io.Writer
+
+interface Action : Serializable {
+    fun perform(path: Path)
+
+    fun perform(writer: Writer)
+}
