+++ app/src/main/kotlin/org/fossify/draw/activities/SimpleActivity.kt
@@ -1,0 +1,30 @@
+package org.fossify.draw.activities
+
+import org.fossify.commons.activities.BaseSimpleActivity
+import org.fossify.draw.R
+
+open class SimpleActivity : BaseSimpleActivity() {
+    override fun getAppIconIDs() = arrayListOf(
+        R.mipmap.ic_launcher_red,
+        R.mipmap.ic_launcher_pink,
+        R.mipmap.ic_launcher_purple,
+        R.mipmap.ic_launcher_deep_purple,
+        R.mipmap.ic_launcher_indigo,
+        R.mipmap.ic_launcher_blue,
+        R.mipmap.ic_launcher_light_blue,
+        R.mipmap.ic_launcher_cyan,
+        R.mipmap.ic_launcher_teal,
+        R.mipmap.ic_launcher_green,
+        R.mipmap.ic_launcher_light_green,
+        R.mipmap.ic_launcher_lime,
+        R.mipmap.ic_launcher_yellow,
+        R.mipmap.ic_launcher_amber,
+        R.mipmap.ic_launcher,
+        R.mipmap.ic_launcher_deep_orange,
+        R.mipmap.ic_launcher_brown,
+        R.mipmap.ic_launcher_blue_grey,
+        R.mipmap.ic_launcher_grey_black
+    )
+
+    override fun getAppLauncherName() = getString(R.string.app_launcher_name)
+}
