+++ app/src/main/kotlin/org/fossify/draw/activities/SplashActivity.kt
@@ -1,0 +1,11 @@
+package org.fossify.draw.activities
+
+import android.content.Intent
+import org.fossify.commons.activities.BaseSplashActivity
+
+class SplashActivity : BaseSplashActivity() {
+    override fun initActivity() {
+        startActivity(Intent(this, MainActivity::class.java))
+        finish()
+    }
+}
