+++ app/src/main/kotlin/org/fossify/draw/extensions/Context.kt
@@ -1,0 +1,6 @@
+package org.fossify.draw.extensions
+
+import android.content.Context
+import org.fossify.draw.helpers.Config
+
+val Context.config: Config get() = Config.newInstance(applicationContext)
