+++ app/src/main/kotlin/org/fossify/draw/interfaces/CanvasListener.kt
@@ -1,0 +1,7 @@
+package org.fossify.draw.interfaces
+
+interface CanvasListener {
+    fun toggleUndoVisibility(visible: Boolean)
+
+    fun toggleRedoVisibility(visible: Boolean)
+}
