+++ app/src/main/res/values/donottranslate.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <string name="package_name">com.simplemobiletools.draw.pro</string>
+    <string name="package_name">org.fossify.draw</string>
 
     <!-- Release notes -->
     <string name="release_38">Added optional zooming, has to be enabled in the app settings</string>
