+++ fastlane/Appfile
@@ -1,2 +1,2 @@
 json_key_file("fastlane/fastlane.json") # Path to the json secret file - Follow https://docs.fastlane.tools/actions/supply/#setup to get one
-package_name("com.simplemobiletools.draw.pro") # e.g. com.krausefx.app
+package_name("org.fossify.draw") # e.g. com.krausefx.app
