+++ gradle/libs.versions.toml
@@ -14,7 +14,7 @@ app-build-minimumSDK = "23"
 app-build-javaVersion = "VERSION_17"
 app-build-kotlinJVMTarget = "17"
 #versioning
-app-version-appId = "com.simplemobiletools.draw.pro"
+app-version-appId = "org.fossify.draw"
 app-version-versionCode = "1"
 app-version-versionName = "1.0.0"
 [libraries]
